WIP on readme for zome
